% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{colorblind_moma_palettes}
\alias{colorblind_moma_palettes}
\title{Names of colorblind-friendly palettes}
\format{
An object of class \code{character} of length 24.
}
\usage{
colorblind_moma_palettes
}
\description{
Lists all palettes that are colorblind-friendly in the package.
To be colorblind-friendly, all colors in the palettes must be distinguishable with deuteranopia, protanopia, and tritanopia.
Use \code{\link{moma.colors}}  to construct palettes or \code{\link{colorblind.friendly.moma}} to test for colorblind-friendliness.
}
\keyword{datasets}
