% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors.R
\docType{data}
\name{MoMAPalettes}
\alias{MoMAPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 31.
}
\usage{
MoMAPalettes
}
\description{
Use names(MoMAPalettes) to return all possible palette names. Current choices are:
\code{Abbott}, \code{Alkalay}, \code{Althoff}, \code{Andri}, \code{Avedon}, \code{Budnitz},
\code{Connors}, \code{Doughton}, \code{Ernest}, \code{Exter}, \code{Flash}, \code{Fritsch},
\code{Klein}, \code{Koons}, \code{Levine1}, \code{Levine2}, \code{Liu}, \code{Lupi},
\code{Ohchi}, \code{OKeeffe}, \code{Palermo}, \code{Panton}, \code{Picabia}, \code{Picasso},
\code{Rattner}, \code{Sidhu}, \code{Smith}, \code{ustwo}, \code{VanGogh}, \code{vonHeyl},
and \code{Warhol}
Use \code{\link{moma.colors}} to construct palettes.
}
\keyword{datasets}
