% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{scale_color_moma_d}
\alias{scale_color_moma_d}
\title{MoMAColors palettes for plotting with ggplot2}
\usage{
scale_color_moma_d(palette_name, direction = 1, override_order = FALSE, ...)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are:
\code{Abbott}, \code{Alkalay1}, \code{Alkalay2}, \code{Althoff}, \code{Andri}, \code{Avedon},
\code{Budnitz}, \code{Clay}, \code{Connors}, \code{Dali}, \code{Doughton}, \code{Ernst},
\code{Exter}, \code{Flash}, \code{Fritsch}, \code{Kippenberger}, \code{Klein}, \code{Koons},
\code{Levine1}, \code{Levine2}, \code{Liu}, \code{Lupi}, \code{Ohchi}, \code{OKeeffe}, \code{Palermo},
\code{Panton}, \code{Picabia}, \code{Picasso}, \code{Rattner}, \code{Sidhu}, \code{Smith}, \code{ustwo},
\code{VanGogh}, \code{vonHeyl}, and \code{Warhol}}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{override_order}{Colors are picked from palette to maximize readability and aesthetics. This means
that colors are not always selected in sequential order from the full palette. If override_order is set to TRUE,
colors are selected in sequential order from the full palette instead. Default is FALSE.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{discrete_scale}}}
}
\description{
Function for using \code{MoMAColors} colors schemes in \code{ggplot2}. Use \code{\link{scale_color_moma_d}} and \code{\link{scale_fill_moma_d}}
for discrete scales and \code{\link{scale_color_moma_c}} and \code{\link{scale_fill_moma_c}} for continuous scales.
}
\examples{
library(ggplot2)
ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
geom_point() +
scale_color_moma_d("Warhol")
}
